+++ Services/Evs/V2/Model/CreateVolumeResponse.cs
@@ -22,7 +22,7 @@ namespace HuaweiCloud.SDK.Evs.V2.Model
         public string JobId { get; set; }
 
         /// <summary>
-        /// 订单ID，云硬盘为包周期计费时返回该参数。 &gt; 说明： &gt;  &gt; - 如果您需要支付订单，请参考：\&quot;[支付包周期产品订单](https://support.huaweicloud.com/api-oce/zh-cn_topic_0075746561.html)\&quot;。
+        /// 订单ID，云硬盘为包周期计费时返回该参数。 &gt; 说明： &gt; 直接在包周期云服务器上新增云硬盘，系统会自动将云硬盘挂载到包周期云服务器上。该情形下也会返回该参数。  &gt; - 如果您需要支付订单，请参考：\&quot;[支付包周期产品订单](https://support.huaweicloud.com/api-oce/zh-cn_topic_0075746561.html)\&quot;。
         /// </summary>
         [JsonProperty("order_id", NullValueHandling = NullValueHandling.Ignore)]
         public string OrderId { get; set; }