+++ Services/FunctionGraph/V2/Model/ListFunctionAsyncInvocationsResult.cs
@@ -1,0 +1,258 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Linq;
+using System.Runtime.Serialization;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using HuaweiCloud.SDK.Core;
+
+namespace HuaweiCloud.SDK.FunctionGraph.V2.Model
+{
+    /// <summary>
+    /// 异步调用记录
+    /// </summary>
+    public class ListFunctionAsyncInvocationsResult 
+    {
+        /// <summary>
+        /// 异步调用状态，支持5种状态 WAIT: 等待 RUNNING: 执行中 SUCCESS: 执行成功 FAIL: 执行失败 DISCARD: 请求丢弃
+        /// </summary>
+        /// <value>异步调用状态，支持5种状态 WAIT: 等待 RUNNING: 执行中 SUCCESS: 执行成功 FAIL: 执行失败 DISCARD: 请求丢弃</value>
+        [JsonConverter(typeof(EnumClassConverter<StatusEnum>))]
+        public class StatusEnum
+        {
+            /// <summary>
+            /// Enum WAIT for value: WAIT
+            /// </summary>
+            public static readonly StatusEnum WAIT = new StatusEnum("WAIT");
+
+            /// <summary>
+            /// Enum RUNNING for value: RUNNING
+            /// </summary>
+            public static readonly StatusEnum RUNNING = new StatusEnum("RUNNING");
+
+            /// <summary>
+            /// Enum SUCCESS for value: SUCCESS
+            /// </summary>
+            public static readonly StatusEnum SUCCESS = new StatusEnum("SUCCESS");
+
+            /// <summary>
+            /// Enum FAIL for value: FAIL
+            /// </summary>
+            public static readonly StatusEnum FAIL = new StatusEnum("FAIL");
+
+            /// <summary>
+            /// Enum DISCARD for value: DISCARD
+            /// </summary>
+            public static readonly StatusEnum DISCARD = new StatusEnum("DISCARD");
+
+            private static readonly Dictionary<string, StatusEnum> StaticFields =
+            new Dictionary<string, StatusEnum>()
+            {
+                { "WAIT", WAIT },
+                { "RUNNING", RUNNING },
+                { "SUCCESS", SUCCESS },
+                { "FAIL", FAIL },
+                { "DISCARD", DISCARD },
+            };
+
+            private string Value;
+
+            public StatusEnum(string value)
+            {
+                Value = value;
+            }
+
+            public static StatusEnum FromValue(string value)
+            {
+                if(value == null){
+                    return null;
+                }
+
+                if (StaticFields.ContainsKey(value))
+                {
+                    return StaticFields[value];
+                }
+
+                return null;
+            }
+
+            public string GetValue()
+            {
+                return Value;
+            }
+
+            public override string ToString()
+            {
+                return $"{Value}";
+            }
+
+            public override int GetHashCode()
+            {
+                return this.Value.GetHashCode();
+            }
+
+            public override bool Equals(object obj)
+            {
+                if (obj == null)
+                {
+                    return false;
+                }
+
+                if (ReferenceEquals(this, obj))
+                {
+                    return true;
+                }
+
+                if (this.Equals(obj as StatusEnum))
+                {
+                    return true;
+                }
+
+                return false;
+            }
+
+            public bool Equals(StatusEnum obj)
+            {
+                if ((object)obj == null)
+                {
+                    return false;
+                }
+                return StringComparer.OrdinalIgnoreCase.Equals(this.Value, obj.Value);
+            }
+
+            public static bool operator ==(StatusEnum a, StatusEnum b)
+            {
+                if (System.Object.ReferenceEquals(a, b))
+                {
+                    return true;
+                }
+
+                if ((object)a == null)
+                {
+                    return false;
+                }
+
+                return a.Equals(b);
+            }
+
+            public static bool operator !=(StatusEnum a, StatusEnum b)
+            {
+                return !(a == b);
+            }
+        }
+
+
+        /// <summary>
+        /// 异步调用请求ID
+        /// </summary>
+        [JsonProperty("request_id", NullValueHandling = NullValueHandling.Ignore)]
+        public string RequestId { get; set; }
+
+        /// <summary>
+        /// 异步调用状态，支持5种状态 WAIT: 等待 RUNNING: 执行中 SUCCESS: 执行成功 FAIL: 执行失败 DISCARD: 请求丢弃
+        /// </summary>
+        [JsonProperty("status", NullValueHandling = NullValueHandling.Ignore)]
+        public StatusEnum Status { get; set; }
+        /// <summary>
+        /// 异步调用错误信息，如果执行成功，则返回空
+        /// </summary>
+        [JsonProperty("error_message", NullValueHandling = NullValueHandling.Ignore)]
+        public string ErrorMessage { get; set; }
+
+        /// <summary>
+        /// 异步调用开始时间（格式为YYYY-MM-DD&#39;T&#39;HH:mm:ss,UTC时间）。
+        /// </summary>
+        [JsonProperty("start_time", NullValueHandling = NullValueHandling.Ignore)]
+        public DateTime? StartTime { get; set; }
+
+        /// <summary>
+        /// 异步调用结束时间（格式为YYYY-MM-DD&#39;T&#39;HH:mm:ss,UTC时间）。
+        /// </summary>
+        [JsonProperty("end_time", NullValueHandling = NullValueHandling.Ignore)]
+        public DateTime? EndTime { get; set; }
+
+
+        /// <summary>
+        /// Get the string
+        /// </summary>
+        public override string ToString()
+        {
+            var sb = new StringBuilder();
+            sb.Append("class ListFunctionAsyncInvocationsResult {\n");
+            sb.Append("  requestId: ").Append(RequestId).Append("\n");
+            sb.Append("  status: ").Append(Status).Append("\n");
+            sb.Append("  errorMessage: ").Append(ErrorMessage).Append("\n");
+            sb.Append("  startTime: ").Append(StartTime).Append("\n");
+            sb.Append("  endTime: ").Append(EndTime).Append("\n");
+            sb.Append("}\n");
+            return sb.ToString();
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public override bool Equals(object input)
+        {
+            return this.Equals(input as ListFunctionAsyncInvocationsResult);
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public bool Equals(ListFunctionAsyncInvocationsResult input)
+        {
+            if (input == null)
+                return false;
+
+            return 
+                (
+                    this.RequestId == input.RequestId ||
+                    (this.RequestId != null &&
+                    this.RequestId.Equals(input.RequestId))
+                ) && 
+                (
+                    this.Status == input.Status ||
+                    (this.Status != null &&
+                    this.Status.Equals(input.Status))
+                ) && 
+                (
+                    this.ErrorMessage == input.ErrorMessage ||
+                    (this.ErrorMessage != null &&
+                    this.ErrorMessage.Equals(input.ErrorMessage))
+                ) && 
+                (
+                    this.StartTime == input.StartTime ||
+                    (this.StartTime != null &&
+                    this.StartTime.Equals(input.StartTime))
+                ) && 
+                (
+                    this.EndTime == input.EndTime ||
+                    (this.EndTime != null &&
+                    this.EndTime.Equals(input.EndTime))
+                );
+        }
+
+        /// <summary>
+        /// Get hash code
+        /// </summary>
+        public override int GetHashCode()
+        {
+            unchecked // Overflow is fine, just wrap
+            {
+                int hashCode = 41;
+                if (this.RequestId != null)
+                    hashCode = hashCode * 59 + this.RequestId.GetHashCode();
+                if (this.Status != null)
+                    hashCode = hashCode * 59 + this.Status.GetHashCode();
+                if (this.ErrorMessage != null)
+                    hashCode = hashCode * 59 + this.ErrorMessage.GetHashCode();
+                if (this.StartTime != null)
+                    hashCode = hashCode * 59 + this.StartTime.GetHashCode();
+                if (this.EndTime != null)
+                    hashCode = hashCode * 59 + this.EndTime.GetHashCode();
+                return hashCode;
+            }
+        }
+    }
+}